Fix some issues with long-running programs and encodings
With OpenSSH, the stdin_text feature relied on the executed program
to read as much input as it needed, end exit. This made it
unusable with long-running programs like nc (netcat).
Fix this by passing the input as escaped string and piping it through
echo.

Based heavily on a patch by Aleksei Slaikovskii:
https://pagure.io/python-pytest-multihost/pull-request/9