drop extra newline from stdin processing
https://pagure.io/python-pytest-multihost/issue/12

* Modified run_command to add -n to echo
* added test_piping_input
* dropped newline from assert for:
  - test_escaping
  - test_escaping_binary
  - test_background_explicit_wait
  - test_background_context

echo -e used by stdin_text processing is appending an extra
newline character to stdin.  For some commands this causes
problems.

$ echo -e test|base64
dGVzdAo=

vs:

$ echo -en test|base64
dGVzdA==

and:

$ echo -en test> test_file
$ cat test_file|base64
dGVzdA==

So if stdin is a string that doesn't already contain a newline,
we shouldn't append it as extra to the end of what we pass to
the command.

Signed-off-by: Scott Poore <spoore@redhat.com>